New Datastore regression test
This test is a ruby version of gcloud-node's datastore test.
Some tests are skipped because they test features not yet implemented.
The index configuration is added to regressions/data, and guidance
is given for building the required indexes in CONTRIBUTING.
Previous regression tests are removed as they are no longer needed.