Refactor Datastore Dataset
Break the old Connection responsibilities into Dataset, Credentials, and Connection.
Dataset is responsible for presenting a nicer API to the users.
Credentials is responsible for hosting the Signet client and signing http requests.
Connection is responsible for making the actual http calls to the API.

This refactor more closely follows the node and python designs, as well as
enables more meaningful unit tests.