Add Properties object
Properties allows for properties to be removed from an entity,
not simply setting the value to NULL.
Properties is a hash-like data structure that does several things for us.
It ensures that only strings are stored as keys, although it allows
symbols to be used for getting and setting values.