Publish messages in batch
The Pub/Sub publish API allows multiple messages to be sent at once.
This allows for more efficent API calls when publishing several messages.
Instead of introducing a single and batch method, use batch when calling
publish with a block. The block is passed a batch object that accepts
all the publish calls to be completed when the block completes.