Update Message's deadline
When pulling messages the subscriber must acknowledge within the deadline.
If the subscriber wants the message to be available to be pulled again
immediately, they can update the deadline to 0. Or, if the subscriber
needs more time, they can update the deadline for the time they need.
This functionality is implemented in Message#delay!