Publish Pubsub messages lazily
Add the ability for Topic#publish to publish from a lazy object.
When a publish fails because the topic does not exist and autocreate is set,
a second HTTP call is made to create the topic, and a third HTTP call is made
to retry the publish.

Fix bug with errors not being properly raised.