Hashify Pub/Sub Message#attributes
Calling a message's attributes should give you a hash, so you
can ask for the keys or iterate over the values.
The code was returning the Google API Client object, which
wraps the hash but does not have all the hash methods.
Convert the object to a hash using to_hash.

Ruby 1.9.3 does not have Hash#to_hash, so guard against that.