Ensure async pull doesn't timeout
It was possible that waiting for a pull request to finish would
lead to an HTTP timeout error. Catch the errors and return an
empty list instead.

Forced to disable rubocop because none of the lines in pull
are at a different level of abstraction. No obvious choice
to extract methods. It would be more confusing to break
the method up.