Allow optional param :project
We want to allow Pub/Sub topic objects to reference topics in
a different project. We allow this by passing in the full path,
but we also want to allow the following syntax:

    pubsub.topic topic-name, project: another-project-id

We implemented this optional parameter when creating full paths
for project, topic, and subscription, although they are not
documented. They are there when/if we need it.