Start implementation of Table Data
Dealing with data is a large effort and will touch many methods in
BigQuery, including Table#data, Job#query_results, and Project#query.
This is the first bit of code dealing with data. More changes coming after.
The data formatting code in this commit should be extracted to objects
responsible for formatting the data in future commits.