Separate View from Table
During regression testing we have noticed that views behave differntly than tables.
Many of the methods on table won't work when it is a view.
Split the classes to avoid potential frustration when things like load and extract
fail. Move query to only be supported on View, since Table has no need for it.