Rename query to query_job
The BigQuery jobs.insert API is different than job.query. The jobs.query
API is simpler, and as of yet unsupported in gcloud. We are going to add
it, but we want to name it query since it should be easier to use. Rename
the current query method to query_job to indicate that it returns a job,
not the data directly.