Add next? and next to Data
We want to make it easier to paginate data on Data and QueryData.
The mechanisms for pagination on Dataa vs. QueryData are different, so we
want to create a shared API so users won't have to care if they have Data
or QueryData. The next? and next methods on Data enable pagination without
calling back to the Table object. This means Data needs to keep a reference
to the Table object. We expect QueryData will need a reference to a
Connection object as well.