Allow NULL values
A Datastore Value is considered NULL if there is no typed value set.
It is possible to have NULL values in Datastore, but it hasn't been possible
to set NULL values from gcloud. Futher, a NULL property value in Datastore
will raise an error when creating an Entity object in gcloud.
Correct this oversight and allow nil to represent NULL in Datastore.

[fixes #259]