Fix bug with string table names
The intent is that the current project and dataset values are used when copying
to a table identified by a string if the string does not contain them.
The bug here is that the nil values were kept when merged.
Removing the keys with nil values before the merge corrects the behavior.

[fixes #326]