Add pages feature task
This task is intended to be used from a feature branch. It will update the docs
in the gh-pages branch from the feature branch. It can optionally push the change
if the push argument is specified. The intended use is to push the updated
gh-pages branch to an individual's github repo, so the docs can be seen without
accepting the feature branch to master.

To automatically push to a non-origin repo, specify the remote:

    $ rake pages:feature[blowmage]