Add Search Index delete
The delete method does not exist on the service, so we are recreating the behavior
in the ruby code. Eventually index will be a resource that can be deleted.
Until then, provide the force option to remove all documents in the index.
This call can be time consuming depending on the number of documents that
exist on an index.