Merge after create
Instead of replacing all document data, merge what was returned.
The Search API is only returning the docId, so index either needs to get
the document to get rank and all the fields, or merge the docId into the
values already on the document object. This is a compromise to keep the
total number of API calls as low as possible.