Remove gRPC memoization
We are seeing many of the gRPC acceptance tests failing. When the gRPC tests run
early they pass, but when they run later they usually fail. We suspect that this
is because of how Logging Service memoizes the gRPC service stub objects.
Remove the memoization of the gRPC objects seems to fix these runtime errors.
Update the logging tests to use a different mechanism to mock the gRPC objects.