Improve Dataset#key convenience method
Allow a key to be created by a path instead of just the kind/id|name combination.
This will allow for fully nested key objects to be created in a single call.
This change is backwards compatible with the existing helper method signature.
Also add optional arguments for setting project and namespace on the keys.