Add Translate service
Add implementation, developer tests, and acceptance tests.
This is a big commit due to the churn in getting authentication correct.
The API discovery doc makes it seem that an OAuth token can be used,
but after several trials and errors this does not seem to be the case.
Authentication must be performed with an API Key, as it is done in
other language implementations such as gcloud-node.