Update recommended params in Vision docs
We want to encourage users to use boolean flags instead of setting max result
values in every method call. A custom max result value can be given, but should
be presented as a secondary option to using the boolean flag.

Update method signatures to prioritize boolean flags as default values.