Add Pub/Sub Subscription::List#next and next?
This is a refactor of Subscription::List, since it is used by both
Project and Topic to list subscriptions. Now that the list holds a
connection object it needs to know which method to call.

Remove prefix optional argument from project#subscriptions. I was a no-op.
The protobuf doesn't have an option for prefix. It must have been a hold over
from the JSON REST API.