Small memory optimization
Don't load the IO object's contents into memory as a string until needed.
This allows the string contents to be garbage collected, which is useful when
users have lots and lots of Vision Image objects created.
Instead, keep references to the IO object.

Also memoize the #to_s value so the Image doesn't keep rewinding the IO object
every time the Image object is inspected.

Change the Base64 encoding to be strict since it shaves off a few bytes as well.