Delay loading GRPC until its used
This change is an attempt to solve issues using the GRPC client in
environments like Rails that fork subprocesses to handle requests.
The GRPC client is initialized when require grpc is called, so we are
attempting to control when that is called.
With this change it is called only when an API call is going to be made.
The protobuf files can still be created, however.