Configure open_timeout_sec for uploads
Uploads use open_timeout_sec instead of timeout_sec because the request uploads data.
Configure open_timeout_sec with timeout so users can configure the client for uploads.
Both timeout_sec and open_timeout_sec are allowed to be nil, but for some reason the
default value for open_timeout_sec is 20. Allow it to be set to nil by default.
When nil, the default timeout_sec seems to be 60, but open_timeout_sec is 120.

[fixes #876]