Add Travis-CI config for SSL certs on OS X
This change is attempting to fix a failure on Ruby 2.0.0 due to missing SSL certificates.
The assumption is that RVM fixed this issue for newer rubies, but not this one.
The error this is attempting to fix is:
Faraday::SSLError: SSL_connect returned=1 errno=0 state=SSLv3 read server certificate B: certificate verify failed