Add googleauth dependency to core
Both google-gax and google-api-client have a dependency on googleauth,
so there is no need for error_reporting or monitoring to also delcare the dependency.
However, google-cloud-core also depends on googleauth, but doesn't have a dependency
on either google-gax or google-api-client.
Make the dependency in google-cloud-core explicit.

We will need to cut new versions of all gems when googleauth revs to 0.6 or higher,
because it is the one dependency shared by all gems.