Add prefix to Datastore acceptance tests
The new CircleCI environment is much faster, so fast that some of these tests are failing now.
Add a prefix to all key names so we don't have tests cleaning up after each other.
Add a transaction to the GoT tests, since that is what the python tests do now.