Update default legacy_sql attribute
Legacy vs. Standard SQL can be set in more ways than this attribute value.
If the query string starts with #standardSQL the query will be standard SQL.
Also, if the user provides query parameters the query will be set to standard.
For these reasons the default value has been changed to nil.
If the value is not set, either by the query comment or use of query parameters
then the query will use legacy SQL.

Rename the attribute to be more in line with the other attribute names.