Add Datastore GRPCUtils
This change has two benefits. First, it ensures that Datastore does not pollute the Core
GRPCUtils module, so it won't behave differently for other services depending on
whether Datastore is loaded or not. Second, it hides Google::Cloud::Core from the docs
so it won't show up in YARD output like rubydocs.info.