Update TIMESTAMP formatting
Add nanoseconds to the timestamp, making it much more accurate.
The TIMESTAMP format seems to have been updated since this code was first written.
However, the new format specified in the documentation generates an error.
For instance, we can't specify time zone offset, so continue to use Z.
Also, add nanoseconds to the previous format, which seems to work fine.