Implement retry for broken streams
Buffer results until a resume_token is received. Once received merge and
return all applicable result rows.
If a stream is broken with ABORTED then retry the query using the
resume_token and clear out all buffered results.

Update the dependency on GRPC so we can use GRPC::Aborted in the code.

These changes are left in an intentionally raw state. I'm not sure what
the right abstraction for this behavior is yet, so the code is ugly.
It is assumed that this code will be refactored in a later commit.

[closes #1236]