Add pairs argument to Spanner Results
The pair method argument allows rows and types to be returned as a
nested Array of pairs rather than a Hash. This will be useful for
when users have rows that have duplicate names. The values can be
represented in a data structure that allows for duplicate keys, while
being familiar to Rubyists.