Update unstable tests
These tests have shown some instability starting with the introduction of the
Pool's thread pool using concurrent-ruby. This is because these tests were
written before then, and all allotment of session and transactions happened
serially. But the thread pool can choose to initialize sessions and transactions
in different order, which is why these tests were failing.

Fix by loosening up the mocks and only checking the number of resources created.