Disable acceptance at_exit hook
Because we are creating a client in the acceptance test helper before the at_exit
hook is invoked by Minitest.autorun, the thread pool is being shutdown prematurely.
Tell Concurrent-Ruby to disable use of their at_exit hook for acceptance tests.
This will allow us to remove use of the caller_runs fallback.