Log entry label values can be dynamic `Proc`
Label values can take the form of either static strings or a
user-defined function (any object that responds to `#call`) which is
passed the current Rack request environment at the moment of generating
the log entry.

The label key remains static and the return value of the function
becomes the label value in the log entry.

The tests show an example of writing labelling log entries with a value
from an HTTP header.

We use this in production to tie together all traces from a user session.