Update Spanner Acceptance Test Setup
We have a limited number of instances available for running tests.
The Spanner team has asked that we reuse a instance instead of
creating and deleting an instance for each test run.
Create a shared instance and use a new database per test run.
Remove the acceptance test that creates instances.