Hand Edits
The hand edits here are made to the generated protobuf files based on my
understanding of how https://github.com/googleapis/artman/issues/288
will be resolved.

To be more explicit: the resolution will allow for the ability to
specify the output namespace of the generated protobuf files. In this
case the data transfer namespace will be
Google::Cloud::Bigquery::DataTransfer::V1. This will correspond to the
following changes:
1) the `Datatransfer` module will be changed to `DataTransfer`
2) the path `datatransfer` will change to `data_transfer`