Allow Job location to be unset
Users may decide to set location to nil in order to unset the value
and make the BigQuery service use the default location.
Currently, this sends the JSON value null to the BigQuery service,
but that will raise the following error:

  invalid: Invalid value for: null is not a valid value

Treat setting location to nil the same as unsetting the value,
this will not send any value for location to the BigQuery service.