Guard against GCS rate limits in BigQuery
This adds a new global method that will guard against rate limit error
when trying to create or delete buckets.
A global method was choosen because it was the simplest solution,
but a more reusable approach can be choosen later.
It is expected that projects that use this method add it to their
acceptance test helper file.