Enable frozen_string_literal magic comment
When memory profiling an app using this gem, I noticed that a large number
(hundreds of thousands )of "@grpc" and "ASCII-8BIT" strings allocated. This
would happen with each message, whic

This patch addresses the issue and future issues by enforcing a frozen string
literals for all files in this library, and adding a rubocop rule to enforce
this pattern in new files.