Firestore: remove unneeded base64 encoding applied to BLOB fields
It looks like fields of type 'bytes' can be set using binary streams. Before the change however the contents of a
stream would be base64-encoded in `google-cloud-firestore/lib/google/cloud/firestore/convert.rb` which would lead to
the actual value of the corresponding 'bytes' field to be base64 encoded as well (in addition to the base64 encoding
the Web Console does, for example).