Add Watch conformance tests
This change supports the newly added conformance tests for Watch.
There are some changes in the Watch implementaton to make these
conformance tests pass.
This includes removal of the code that allowed the insertion of
documents into the red-black tree in a more performant way,
and sorting all added documents on initial insert.