Specify YARD Markdown provider (#2232)
All gems use Markdown syntax for documentation, and we have configured
YARD to generate documentation accordingly. However, we also use some
Markdown options, such as fenced code blocks, that are not standard
Markdown and not supported in all Markdown providers. Set the Markdown
provider to redcarpet to ensure that documentation generates properly.

* Specify redcarpet in the .yardopts file.
* Add redcarpet as a development dependency. This will ensure that
  users who generate documentation from the gem will produce
  properly formatted documentation.
* Add YARD as a development dependency to gems that were missing it.