Add docs tasks and YardBuilder
These tasks will live alongside the jsondoc tasks for now.
Eventually, we will stop using jsondoc and use these docs.
YardBuilder is the code that generates the YARD documentation
using the template on the gh-pages branch.
The generated documentation files will be located in the
gh-pages branch.
New releases will be added to the _data/releases.yaml file
in the gh-pages branch.