Refactor Subscriber Stream shutdown (#2278)
Change the shutdown process to no longer process pending callbacks. When the
number of pending callbacks was too high the thread pool would hang. This new
behavior also seems to be more expected.

Change the wait behavior to only wait on pending change requests. Limit waiting
on pending change requests, so the Subscriber does not block indefinitely.

We can make this change because we are using the Unary pusher and not the Stream
pusher, because the Unary pusher does not require that we keep the stream open
to send change requests.