Update documentation files
Add a new main page for the YARD documentation that contains the "guide" content,
and supports YARD links using {}, which the README.md file cannot, since it is
intended to be the main page shown on GitHub.
Add the other files, CODE_OF_CONDUCT, CONTRIBUTING, FAQ, TROUBLESHOOTING, but enable
YARD links on them, since their intended use is in documentation, not on GitHub
as is the top-level version of these files.
Customize these new files to be specific to gem specific, instead of project-wide.