Update partial download test
This test fails occasionally on CI. We aren't sure why.
We have a custom message to help dignose, but we have seen some
download verification errors when it gets proc'd.
Set the download to verify: :none to avoid the verification errors.
Update the custom message to show a bit more information.