Set Debugger evaluation_time_limit during coverage
The value that was set previously by test/helper.rb would be reset
by some of the other tests that reset configure as part of the test.
Set a value in the test attribute getter method by looking at an ENV VAR.
This will ensure that all tests use the increased value during coverage.