Set evaluation_time_limit in Evaluator tests
This extends a fix to the Debugger tests when running in the coverage task.
We see frequent failures on CI because debugger tests fail during coverage.
We think this is due to the increased overhead by the coverage reporting.
We tried to extend this time limit before, but it was being reset by
some of the other tests.
Add the time limit extension to the Evaluator tests, which do not share the
same test setup as most of the other Debugger tests.