Make Logger thread-safe
Make sure all calls accessing the @request_info_map Hash object are
using the synchonize block.
Rename @request_info to @request_info_map, because this hash is not a
RequestInfo object, but a map of RequestInfo objects for threads.
Renamed so it was more clear that this object is a repository,
not an actual RequestInfo object.