Add buffering and batches to AsyncWriter
Update AsyncWriter to buffer log entries.
The previous implementation was making API requests
as fast as it could. This change allows the writer
to buffer the entries longer and send in more batches.
Also add configuration for the max number of entries
to buffer, and well as the max byte size to buffer
before the batch is sent.

Deprecate suspend and resume, which are less useful now.