Switch to CachedThreadPool
This will allow the AsyncWriter add threads to the thread pool as needed,
instead of allocating all the threads up-front. This will hopefully
reduce the resources allocated until they are needed.

For instance, if a AsyncWriter never needs more than 3 threads, even under
heavy usage, then that is all the threads that will be allocated.