Add default insert_id value for Entry
An Entry object is assigned an insert_id when created so that if the
Entry object gets persisted multiple times it would know its insert_id
value and not attempt to generate a new one for each persist attempt.
An Entry object will still be considered empty if the only value it has
is the insert_id.