Move @new_sessions_in_progress usage
Move all modifications of this variable to the new_session! method.
The intention is to limit the places that this can change.
new_session! increments the count, but does not raise SessionLimitError.
The responsibility for raising if the count is higher than the limit
is still on the method calling new_session!.