Add Ordering Keys
These new feature add the ability to ensure related messages are published in
the order they were provided to the client. And to ensure related messages are
sent to subscribers in the order in which they were published. Messages can be
tagged with an ordering key, a string that identifies related messages for which
publish order should be respected.