Add `File.open` to quickstart example
Currently, the `response` object in the example is only stored in memory. It's not self-evident that the `response`'s `audio_content` method is the way to work with the audio.  Since the example is encoded as an MP3, I think it's natural to show the first-time user how to save that file to disk, and introduce them to that method.