Fix coverage tasks
The coverage tasks were not reporting correctly due to some
changes in how the test task is being defined.
The require of the SimpleCov library now has to be declared
in the test files, before Minitest is required.
The coverage configuration is moved to a .simplecov file.
The SimpleCov dependency has been updated.
The top-level coverage task's configuration of files to
ignore in the report has been simplified and updated.