doc: document single use as default
Document that Cloud Spanner will use a single use read-only transaction
as default for queries when no other option is specified. Also changes
the example code for the database client to show how to execute a single
query using a single use read-only transaction, instead of a read/write
transaction.