Determine ALTER TABLE to be a DDL statement
Semi-related to comments on https://github.com/googleapis/google-cloud-ruby/issues/9617#issuecomment-784501338.

We found that re-running our schema migrations (after deleting the dataset) (using sequel-bigquery) sometimes resulted in the following error inexplicably complaining that the dataset could not be found:

```
Google::Cloud::NotFoundError: notFound: Not found: Dataset greensync-dex-dev:local_dev_ZimbiX
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/google-cloud-bigquery-1.37.0/lib/google/cloud/bigquery/service.rb:572:in `rescue in execute'
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/google-cloud-bigquery-1.37.0/lib/google/cloud/bigquery/service.rb:565:in `execute'
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/google-cloud-bigquery-1.37.0/lib/google/cloud/bigquery/service.rb:224:in `list_tabledata'
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/google-cloud-bigquery-1.37.0/lib/google/cloud/bigquery/query_job.rb:735:in `data'
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/google-cloud-bigquery-1.37.0/lib/google/cloud/bigquery/dataset.rb:1709:in `query'
/home/brendan/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/sequel-bigquery-0.6.0/lib/sequel-bigquery.rb:85:in `block (2 levels) in execute'
```

The dataset definitely did exist (due to previous migration files succeeding), so this was quite a confusing error. Drilling into the backtrace, we discovered that `QueryJob#data` has a special return path for DDL statements; yet the `#ddl?` method did not categorise an `ALTER TABLE` statement as DDL!

We have no idea why `service.list_tabledata` was failing inconsistently, but after making the change to determine `ALTER TABLE` to be a DDL statement, we have not seen this failure after many runs.

Oddly, the method is defined twice (one does not seem to be used?), so we made the change to both.

Although we discovered this issue in 1.37.0, the applicable code is unchanged in the latest version, 1.38.0.

Co-Authored-By: Rob Harrington <rob.harrington@greensync.com.au>