Fix Firestore::Convert.fields_to_hash
Because `fields` is an instance of `Google::Protobuf::Map` and not a
`Hash`, the `#to_h` method ignores the given block, and the value
returned instead of something like:

    {id: 123}

looke like:

   {"id"=>{:null_value=>:NULL_VALUE, :boolean_value=>false, :integer_value=>123, :double_value=>0.0, :timestamp_value=>nil, :string_value=>"", :bytes_value=>"", :reference_value=>"", :geo_point_value=>nil, :array_value=>nil, :map_value=>nil}

The existing tests already caught the issue, so no new tests were
neccessary.

The issue was first introduced [here][1], and affects version `2.7.0` of the gem.

[1]: https://github.com/googleapis/google-cloud-ruby/pull/18446/files#diff-2b92cbc5743990a07fd45c459abcd96b9b4673566be07698bde11b62ba674f07L53