chore(perf): Make all gems load faster
Gem.find_latest_files's performance scales with the amount of files contained
across all gems of a project. This can quickly get out of hand,
especially on projects with many gems, or large gems.

Gem.loaded_spec is already populated by bundler, and we can iterate over
only the name of the gems to detect those matching the google-cloud-*
pattern. We can then guess the name of the file we expect to be present.

On my machine, this makes `auto_load_files` go from ~100ms to <1ms, and
requiring `bigtable` goes from ~220ms to ~110.